Added lab03 solutions
There are two possible solutions I implemented: One is the optimal that passes all the test cases and runs a modified Dijkstra algorithm, but it prioritizes the nodes
with the smallest ks and then sorts by current path length. This means that the first k we find that satisfies the condition is actually the minimum possible, which is
what the initial question asks. The second solution uses the same Dijkstra approach, with the difference that it sorts the priority queue by path length and then searches
for the minimal k with a binary search. This is possible as the dp[node][k] array will always be sorted as far as k is concerned, meaning dp[node][k + 1] <= dp[node][k],
for every k.